+++ prototypes/item/items.lua
@@ -12,7 +12,7 @@ if DECT.ENABLED["landscaping"] then
 		},
 		{
 			type = "item",
-			name = "base-dirt",
+			name = "dect-base-dirt",
 			icon = "__Dectorio__/graphics/icons/base-dirt.png",
 		    flags = {"goes-to-main-inventory"},
 		    subgroup = "landscaping",
@@ -26,7 +26,7 @@ if DECT.ENABLED["landscaping"] then
 		},
 		{
 			type = "item",
-			name = "base-sand",
+			name = "dect-base-sand",
 			icon = "__Dectorio__/graphics/icons/base-sand.png",
 		    flags = {"goes-to-main-inventory"},
 		    subgroup = "landscaping",
@@ -40,7 +40,7 @@ if DECT.ENABLED["landscaping"] then
 		},
 		{
 			type = "item",
-			name = "base-grass",
+			name = "dect-base-grass",
 			icon = "__Dectorio__/graphics/icons/base-grass.png",
 		    flags = {"goes-to-main-inventory"},
 		    subgroup = "landscaping",
@@ -69,56 +69,56 @@ if DECT.ENABLED["painted-concrete"] then
 		},
 		{
 			type = "item",
-			name = "paint-hazard",
+			name = "dect-paint-hazard",
 			icon = "__Dectorio__/graphics/icons/paint-hazard.png",
 		    flags = {"goes-to-main-inventory"},
 		    subgroup = "concrete-paint",
 		    order = "a[paint-hazard]",
 		    stack_size = 100,
 		    place_as_tile = {
-				result = "paint-hazard-left",
+				result = "dect-paint-hazard-left",
 				condition_size = 4,
 				condition = { "water-tile" }
 			}
 		},
 		{
 			type = "item",
-			name = "paint-emergency",
+			name = "dect-paint-emergency",
 			icon = "__Dectorio__/graphics/icons/paint-emergency.png",
 		    flags = {"goes-to-main-inventory"},
 		    subgroup = "concrete-paint",
 		    order = "a[paint-hazard]-b[paint-emergency]",
 		    stack_size = 100,
 		    place_as_tile = {
-				result = "paint-emergency-left",
+				result = "dect-paint-emergency-left",
 				condition_size = 4,
 				condition = { "water-tile" }
 			}
 		},
 		{
 			type = "item",
-			name = "paint-safety",
+			name = "dect-paint-safety",
 			icon = "__Dectorio__/graphics/icons/paint-safety.png",
 		    flags = {"goes-to-main-inventory"},
 		    subgroup = "concrete-paint",
 		    order = "a[paint-hazard]-b[paint-emergency]-c[paint-radiation]-d[paint-safety]",
 		    stack_size = 100,
 		    place_as_tile = {
-				result = "paint-safety-left",
+				result = "dect-paint-safety-left",
 				condition_size = 4,
 				condition = { "water-tile" }
 			}
 		},
 		{
 			type = "item",
-			name = "paint-radiation",
+			name = "dect-paint-radiation",
 			icon = "__Dectorio__/graphics/icons/paint-radiation.png",
 		    flags = {"goes-to-main-inventory"},
 		    subgroup = "concrete-paint",
 		    order = "a[paint-hazard]-b[paint-emergency]-c[paint-radiation]",
 		    stack_size = 100,
 		    place_as_tile = {
-				result = "paint-radiation-left",
+				result = "dect-paint-radiation-left",
 				condition_size = 4,
 				condition = { "water-tile" }
 			}
@@ -132,14 +132,14 @@ if DECT.ENABLED["wood-floor"] then
 	data:extend({
 		{
 			type = "item",
-			name = "wood-floor",
+			name = "dect-wood-floor",
 			icon = "__Dectorio__/graphics/icons/wood-floor.png",
 		    flags = {"goes-to-main-inventory"},
 		    subgroup = "terrain",
 		    order = "a[stone-brick]-b[wood-floor]",
 		    stack_size = 100,
 		    place_as_tile = {
-				result = "wood-floor",
+				result = "dect-wood-floor",
 				condition_size = 4,
 				condition = { "water-tile" }
 			}
@@ -153,14 +153,17 @@ if DECT.ENABLED["walls"] then
 	data:extend({
 		{
 			type = "item",
-			name = "concrete-wall",
+			name = "dect-concrete-wall",
 			icon = "__Dectorio__/graphics/icons/concrete-wall.png",
 		    flags = {"goes-to-quickbar"},
 		    subgroup = "defensive-structure",
 		    order = "a[stone-wall]-b[concrete-wall]",
 		    stack_size = 100,
-	    	place_result = "concrete-wall"
+	    	place_result = "dect-concrete-wall"
 		}
 	})
+
+	-- Change stone wall icon
+	data.raw["item"]["stone-wall"].icon = "__Dectorio__/graphics/icons/stone-brick-wall.png"
 
 end